updated envSetup && envSetupTest

+++ buffet-api/src/Utils/EnvSetup.php
@@ -10,7 +10,8 @@ class EnvSetup
     public string $envPath;
 
     public string $envContent = "DECRYPT_KEY=a_key";
-    protected bool $returnOriginalEnv = false;
+    public bool $returnOriginalEnv = false;
+    public bool $isDummy = false;
 
     public function __construct()
     {
@@ -23,26 +24,32 @@ class EnvSetup
      */
     public function setupDummyEnv($fileContent = ""): void
     {
+        // takes default content if custom content is not provided
         if (empty($fileContent)) {
             $fileContent = $this->envContent;
         }
-        if (file_exists($this->envPath)) {
+
+        if (file_exists($this->envPath) && $this->returnOriginalEnv == false) {
             $this->returnOriginalEnv = true;
             copy($this->envPath, __DIR__ . "/temp.env");
-        } else {
-            if ($env = fopen($this->envPath, "w")) {
-                fwrite($env, $fileContent);
-                fclose($env);
-            }
+        }
+        if ($env = fopen($this->envPath, "w")) {
+            fwrite($env, $fileContent);
+            fclose($env);
+            $this->isDummy = true;
         }
     }
 
     public function cleanupDummyEnv(): void
     {
-        unlink($this->envPath);
+        if ($this->isDummy) {
+            unlink($this->envPath);
+            $this->isDummy = false;
+        }
         if ($this->returnOriginalEnv) {
             copy(__DIR__ . "/temp.env", $this->envPath);
             unlink(__DIR__ . "/temp.env");
+            $this->returnOriginalEnv = false;
         }
     }
